refactors all docker builds out of terraform

+++ terraform/deploy.sh
@@ -3,12 +3,20 @@
 # ADMIN_ARN is set in the ci node env and should not be included in this deploy script
 
 # variables that will likely be changed frequently
-CALCLOUD_VER="0.4.19"
+CALCLOUD_VER="0.4.20"
 CALDP_VER="0.2.11"
 CAL_BASE_IMAGE="stsci/hst-pipeline:CALDP_20210505_CAL_final"
 
 # this is the tag that the image will have in AWS ECR
 CALDP_IMAGE_TAG="latest"
+
+# these variables are overrides for developers that allow the deploy script to build from local calcloud/caldp source
+# i.e. CALCLOUD_BUILD_DIR="$HOME/deployer/calcloud"
+CALCLOUD_BUILD_DIR="$HOME/bhayden/calcloud"
+CALDP_BUILD_DIR=""
+
+#uncomment this to deploy to a custom env name
+# aws_env="your-env-name-here"
 
 # turn CAL_BASE_IMAGE into CSYS_VER by splitting at the :, splitting again by underscore and keeping the
 # first two fields, and then converting to lowercase
@@ -18,28 +26,38 @@ CSYS_VER=`echo $CSYS_VER | awk '{print tolower($0)}'`
 
 # variables that will be changed less-frequently
 TMP_INSTALL_DIR="/tmp/calcloud_install"
-
 mkdir $TMP_INSTALL_DIR
 cd $TMP_INSTALL_DIR
 
-# calcloud source download/unpack
-wget "https://github.com/spacetelescope/calcloud/archive/v$CALCLOUD_VER.tar.gz"
-tar -xvzf "v$CALCLOUD_VER.tar.gz"
-rm "v$CALCLOUD_VER.tar.gz"
+# setting up the calcloud source dir if it needs downloaded
+# equivalent to "if len($var) == 0"
+if [ -z "$CALCLOUD_BUILD_DIR" ]
+then
+    CALCLOUD_BUILD_DIR="${TMP_INSTALL_DIR}/v$CALCLOUD_VER"
+    # calcloud source download/unpack
+    cd $TMP_INSTALL_DIR
+    wget "https://github.com/spacetelescope/calcloud/archive/v$CALCLOUD_VER.tar.gz"
+    tar -xvzf "v$CALCLOUD_VER.tar.gz"
+    rm "v$CALCLOUD_VER.tar.gz"  
+fi
 
-# caldp source download/unpack
-# github's tarballs don't work with pip install, so we have to clone and checkout the tag
-git clone https://github.com/spacetelescope/caldp.git
-cd caldp && git fetch --all --tags && git checkout tags/v${CALDP_VER} && cd ..
+# setting up the caldp source dir if it needs downloaded
+# equivalent to "if len($var) == 0"
+if [ -z "$CALDP_BUILD_DIR"]
+then
+    CALDP_BUILD_DIR="${TMP_INSTALL_DIR}/caldp"
+    cd $CALDP_BUILD_DIR
+    # caldp source download/unpack
+    # github's tarballs don't work with pip install, so we have to clone and checkout the tag
+    git clone https://github.com/spacetelescope/caldp.git
+    cd caldp && git fetch --all --tags && git checkout tags/v${CALDP_VER} && cd ..
+fi
 
 # get a couple of things from AWS ssm
 # the env, i.e. sb,dev,test,prod
 aws_env_response=`awsudo $ADMIN_ARN aws ssm get-parameter --name "environment" | grep "Value"`
 aws_env=${aws_env_response##*:}
 aws_env=`echo $aws_env | tr -d '",'`
-
-#uncomment this to deploy to a custom env name
-# aws_env="your-env-name-here"
 
 # the tf state bucket name
 aws_tfstate_response=`awsudo $ADMIN_ARN aws ssm get-parameter --name "/s3/tfstate" | grep "Value"`
@@ -48,7 +66,7 @@ aws_tfstate=`echo $aws_tfstate | tr -d '",'`
 echo $aws_tfstate
 
 # initial terraform setup
-cd calcloud-${CALCLOUD_VER}/terraform
+cd ${CALCLOUD_BUILD_DIR}/terraform
 
 # terraform init and s3 state backend config
 awsudo $ADMIN_ARN terraform init -backend-config="bucket=${aws_tfstate}" -backend-config="key=calcloud/${aws_env}.tfstate" -backend-config="region=us-east-1"
@@ -61,25 +79,36 @@ repo_url=${repo_url_response##*=}
 # removes double quotes from variable
 repo_url=`echo $repo_url | tr -d '"'`
 
-# build and deploy caldp docker image
-cd ../../caldp
-# cd ~/bhayden/caldp
+##### DOCKER IMAGE BUILDING #########
 CALDP_DOCKER_IMAGE="${repo_url}:${CALDP_IMAGE_TAG}"
-docker build -f Dockerfile -t ${CALDP_DOCKER_IMAGE} --build-arg CAL_BASE_IMAGE=${CAL_BASE_IMAGE}  .
-# need to "log in" to ecr to push the image
+MODEL_DOCKER_IMAGE="${repo_url}:model"
+TRAINING_DOCKER_IMAGE="${repo_url}:training"
+
+# need to "log in" to ecr to push or pull the images
 awsudo $ADMIN_ARN aws ecr get-login-password --region us-east-1 | docker login --username AWS --password-stdin $repo_url
+
+# naming is confusing here but "modeling" directory plus "training" image is correct
+cd ${CALCLOUD_BUILD_DIR}/modeling
+docker build -f Dockerfile -t ${TRAINING_DOCKER_IMAGE} .
+docker push ${TRAINING_DOCKER_IMAGE}
+# jobPredict lambda env
+cd ${CALCLOUD_BUILD_DIR}/lambda/JobPredict
+docker build -f Dockerfile -t ${MODEL_DOCKER_IMAGE} .
+docker push ${MODEL_DOCKER_IMAGE}
+# caldp image
+cd ${CALDP_BUILD_DIR}
+docker build -f Dockerfile -t ${CALDP_DOCKER_IMAGE} --build-arg CAL_BASE_IMAGE=${CAL_BASE_IMAGE}  .
 docker push ${CALDP_DOCKER_IMAGE}
 
-# deploy rest of terraform
-cd ../calcloud-${CALCLOUD_VER}/terraform
+#### PRIMARY TERRAFORM BUILD #####
+cd ${CALCLOUD_BUILD_DIR}/terraform
+
 # must taint the compute env to be safe about launch template handling. see comments in batch.tf
 awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.compute_env[0]
 awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.compute_env[1]
 awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.compute_env[2]
 awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.compute_env[3]
-
-awsudo $ADMIN_ARN terraform taint docker_registry_image.calcloud_predict_model
-awsudo $ADMIN_ARN terraform taint module.lambda_function_container_image.aws_lambda_function.this[0]
+awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.model_compute_env[0]
 
 # manual confirmation required
 awsudo $ADMIN_ARN terraform apply -var "awsysver=${CALCLOUD_VER}" -var "awsdpver=${CALDP_VER}" -var "csys_ver=${CSYS_VER}" -var "environment=${aws_env}"
